api/requirements.txt
asgiref==3.4.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.2.14
django-environ==0.9.0
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-http-client==3.3.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.6.0
